+++ tools/src/main/java/org/apache/kafka/tools/VerifiableProducer.java
@@ -34,6 +34,7 @@ import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.serialization.StringSerializer;
+import org.apache.kafka.common.utils.ThroughputThrottler;
 
 import java.io.IOException;
 import java.io.InputStream;
